% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_storr.R
\name{get_object_storr}
\alias{get_object_storr}
\title{get an object_storr}
\usage{
get_object_storr(path)
}
\arguments{
\item{path}{the local path where object_storr will be created}
}
\value{
The function returns a storr like object (a list) with less functions and a node named "storr_details" which contains two storr_rds objects used for object_storr
}
\description{
get an object_storr
}
\examples{
st <- get_object_storr("test")
st$set(sin, cos)
st$list()
st$get(sin)(pi)
st$destroy()


}
\seealso{
\link[storr]{storr_rds}
}
